+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - pylibmc is no longer a dependency
 - Default cache prefix is now just "pyappcache" and slashes are added when building raw keys
 - The required API surface of caches has been reduced further
+- MemcacheCache will retry exactly once when pylibmc raises a ConnectionError
+  - in order to be robust against memcache restarts
 
 ### Removed
 - repr no longer defined on GenericStringKey